Refactored tests not to be end-to-end
